refactor: 라우트 컴포넌트 분리 (#23)
